Changed packet to use floats for sensor data

+++ firmware/main/packet.h
@@ -6,7 +6,7 @@
 
 typedef struct {
     uint8_t header;
-    uint16_t moisture, light, temperature, humidity;
+    float moisture, light, temperature, humidity;
 } sensor_data_packet_t;
 
 int packet_to_json(char* buffer, size_t buffer_size, const sensor_data_packet_t* packet);
